src/main/webapp/app/entities/printing-centre-sig/detail/printing-centre-sig-detail.component.ts
import { Component, OnInit } from '@angular/core';
import { ActivatedRoute } from '@angular/router';

import { IPrintingCentreSig } from '../printing-centre-sig.model';
import { DataUtils } from 'app/core/util/data-util.service';

@Component({
  selector: 'sigma-printing-centre-sig-detail',
  templateUrl: './printing-centre-sig-detail.component.html',
})
export class PrintingCentreSigDetailComponent implements OnInit {
  printingCentre: IPrintingCentreSig | null = null;

  constructor(protected dataUtils: DataUtils, protected activatedRoute: ActivatedRoute) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ printingCentre }) => {
      this.printingCentre = printingCentre;
    });
  }

  byteSize(base64String: string): string {
    return this.dataUtils.byteSize(base64String);
  }

  openFile(base64String: string, contentType: string | null | undefined): void {
    this.dataUtils.openFile(base64String, contentType);
  }

  previousState(): void {
    window.history.back();
  }
}
